Fix italic bug.
Actually, I think there was a bug in the original code, which the refactoring exposed. It had been a bit of red flag for me that slant/deSlant feel like operations that should happen in matched pairs, but in setSpace there was a deSlant without a corresponding slant.

What has happening was that processMargins was actually modifying the content of the margins array, as well as returning a computed value. Because we are now passing the margins by value instead of by reference (as part of an array), the refactored version doesn't modify the margins any more. But I don't think it ever should have done, and the unmatched deSlant was just compensating for this modification. Now the modification isn't being done, the unmatched deSlant is no longer necessary.